backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "openai>=1.57.0,<2.0.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "redis>=7.2.0",
    "sse-starlette>=3.2.0",
    "tavily-python>=0.7.21",
    "uvicorn[standard]>=0.41.0",
    "httpx==0.27.2",
]